+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 * New: Now possible to control the container name for .NET service using
   a `bake.yaml` file
 * New: `DOTNET_ROLL_FORWARD` now defaults to `LatestMajor`
+* New: Add `-p:PublishReadyToRun=true` for .NET when publishing to specific runtime
 * Fixed: Container paths that contain `/` are now valid
 
 # 0.18-beta